requirements.txt
torch
transformers
accelerate
pillow
sentencepiece
torchvision

